# swagger-spec.json
{"openapi":"3.0.0","paths":{"/api/logs":{"get":{"operationId":"LogsController_get","parameters":[{"name":"page","required":false,"in":"query","schema":{"example":1,"type":"number"}},{"name":"pageSize","required":false,"in":"query","schema":{"example":10,"type":"number"}},{"name":"orderBy","required":false,"in":"query","schema":{"example":"createdAt","type":"string"}},{"name":"sortOrder","required":false,"in":"query","schema":{"example":"DESC|ASC","type":"string"}},{"name":"lang","required":false,"in":"query","description":"The language for the request","schema":{"default":"en","example":"fr","enum":["fr","en","ar"],"type":"string"}},{"in":"header","required":true,"name":"Accept-Language","schema":{"type":"string","default":"fr","enum":["fr","en","ar"]}}],"responses":{"200":{"description":""}},"tags":["Logs"]}},"/api/voucher":{"get":{"operationId":"VoucherController_getAll","parameters":[{"name":"lang","required":false,"in":"query","description":"The language for the request","schema":{"default":"en","example":"fr","enum":["fr","en","ar"],"type":"string"}},{"name":"transactionId","required":false,"in":"query","schema":{"example":"","type":"uuid"}},{"name":"documentId","required":false,"in":"query","schema":{"example":"","type":"uuid"}},{"name":"productId","required":false,"in":"query","schema":{"example":"","type":"uuid"}},{"name":"status","required":false,"in":"query","schema":{"example":1,"type":"number"}},{"name":"page","required":false,"in":"query","schema":{"example":1,"type":"number"}},{"name":"pageSize","required":false,"in":"query","schema":{"example":10,"type":"number"}},{"name":"orderBy","required":false,"in":"query","schema":{"example":"createdAt","type":"string"}},{"name":"sortOrder","required":false,"in":"query","schema":{"example":"DESC","type":"string"}},{"in":"header","required":true,"name":"Accept-Language","schema":{"type":"string","default":"fr","enum":["fr","en","ar"]}}],"responses":{"200":{"description":""}},"tags":["Voucher"]}},"/stock/upload":{"post":{"operationId":"StockController_uploadFile","parameters":[{"in":"header","required":true,"name":"Accept-Language","schema":{"type":"string","default":"fr","enum":["fr","en","ar"]}}],"requestBody":{"required":true,"description":"Upload CSV file to create customers","content":{"multipart/form-data":{"schema":{"type":"object","properties":{"file":{"type":"string","format":"binary"}}}}}},"responses":{"201":{"description":""}}}},"/api/product":{"post":{"operationId":"ProductController_create","parameters":[{"in":"header","required":true,"name":"Accept-Language","schema":{"type":"string","default":"fr","enum":["fr","en","ar"]}}],"requestBody":{"required":true,"description":"Json structure for Product object","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateProductDto"}}}},"responses":{"201":{"description":""}},"tags":["Product"]},"get":{"operationId":"ProductController_getAll","parameters":[{"name":"lang","required":false,"in":"query","description":"The language for the request","schema":{"default":"en","example":"fr","enum":["fr","en","ar"],"type":"string"}},{"name":"productAccountingCode","required":false,"in":"query","schema":{"example":"1","type":"string"}},{"name":"facialId","required":false,"in":"query","schema":{"example":"....","type":"uuid"}},{"name":"operatorId","required":false,"in":"query","schema":{"example":"......","type":"uuid"}},{"name":"enabled","required":false,"in":"query","schema":{"example":true,"type":"boolean"}},{"name":"page","required":false,"in":"query","schema":{"example":1,"type":"number"}},{"name":"pageSize","required":false,"in":"query","schema":{"example":10,"type":"number"}},{"name":"orderBy","required":false,"in":"query","schema":{"example":"createdAt","type":"string"}},{"name":"sortOrder","required":false,"in":"query","schema":{"example":"DESC|ASC","type":"string"}},{"in":"header","required":true,"name":"Accept-Language","schema":{"type":"string","default":"fr","enum":["fr","en","ar"]}}],"responses":{"200":{"description":""}},"tags":["Product"]},"put":{"operationId":"ProductController_update","parameters":[{"in":"header","required":true,"name":"Accept-Language","schema":{"type":"string","default":"fr","enum":["fr","en","ar"]}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateProductDtoTcp"}}}},"responses":{"200":{"description":""}},"tags":["Product"]}},"/api/product/{id}":{"get":{"operationId":"ProductController_getOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}},{"in":"header","required":true,"name":"Accept-Language","schema":{"type":"string","default":"fr","enum":["fr","en","ar"]}}],"responses":{"200":{"description":""}},"tags":["Product"]}},"/api/facial":{"post":{"operationId":"FacialController_create","parameters":[{"in":"header","required":true,"name":"Accept-Language","schema":{"type":"string","default":"fr","enum":["fr","en","ar"]}}],"requestBody":{"required":true,"description":"Json structure for Facial object","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateFacialDto"}}}},"responses":{"201":{"description":""}},"tags":["Facial"]},"put":{"operationId":"FacialController_update","parameters":[{"in":"header","required":true,"name":"Accept-Language","schema":{"type":"string","default":"fr","enum":["fr","en","ar"]}}],"requestBody":{"required":true,"description":"Json structure for Facial object","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateFacialDto"}}}},"responses":{"200":{"description":""}},"tags":["Facial"]},"get":{"operationId":"FacialController_getAll","parameters":[{"name":"page","required":false,"in":"query","schema":{"example":1,"type":"number"}},{"name":"pageSize","required":false,"in":"query","schema":{"example":10,"type":"number"}},{"name":"orderBy","required":false,"in":"query","schema":{"example":"createdAt","type":"string"}},{"name":"sortOrder","required":false,"in":"query","schema":{"example":"DESC|ASC","type":"string"}},{"name":"lang","required":false,"in":"query","description":"The language for the request","schema":{"default":"en","example":"fr","enum":["fr","en","ar"],"type":"string"}},{"name":"label","required":false,"in":"query","schema":{"example":"5 TND","type":"string"}},{"name":"value","required":false,"in":"query","schema":{"example":"1.7","type":"number"}},{"name":"enabled","required":false,"in":"query","schema":{"example":true,"type":"boolean"}},{"in":"header","required":true,"name":"Accept-Language","schema":{"type":"string","default":"fr","enum":["fr","en","ar"]}}],"responses":{"200":{"description":""}},"tags":["Facial"]}},"/api/facial/{id}":{"get":{"operationId":"FacialController_getOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}},{"in":"header","required":true,"name":"Accept-Language","schema":{"type":"string","default":"fr","enum":["fr","en","ar"]}}],"responses":{"200":{"description":""}},"tags":["Facial"]}},"/api/operator":{"post":{"operationId":"OperatorController_create","parameters":[{"in":"header","required":true,"name":"Accept-Language","schema":{"type":"string","default":"fr","enum":["fr","en","ar"]}}],"requestBody":{"required":true,"description":"Json structure for operator object","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateOperatorDto"}}}},"responses":{"201":{"description":""}},"tags":["Operator"]},"get":{"operationId":"OperatorController_getAll","parameters":[{"name":"page","required":false,"in":"query","schema":{"example":1,"type":"number"}},{"name":"pageSize","required":false,"in":"query","schema":{"example":10,"type":"number"}},{"name":"orderBy","required":false,"in":"query","schema":{"example":"createdAt","type":"string"}},{"name":"sortOrder","required":false,"in":"query","schema":{"example":"DESC|ASC","type":"string"}},{"name":"lang","required":false,"in":"query","description":"The language for the request","schema":{"default":"en","example":"fr","enum":["fr","en","ar"],"type":"string"}},{"name":"label","required":false,"in":"query","schema":{"example":"OOREDOO","type":"string"}},{"name":"enabled","required":false,"in":"query","schema":{"example":true,"type":"boolean"}},{"in":"header","required":true,"name":"Accept-Language","schema":{"type":"string","default":"fr","enum":["fr","en","ar"]}}],"responses":{"200":{"description":""}},"tags":["Operator"]}},"/api/operator/{id}":{"put":{"operationId":"OperatorController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}},{"in":"header","required":true,"name":"Accept-Language","schema":{"type":"string","default":"fr","enum":["fr","en","ar"]}}],"requestBody":{"required":true,"description":"Json structure for operator object","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateOperatorDto"}}}},"responses":{"200":{"description":""}},"tags":["Operator"]},"get":{"operationId":"OperatorController_getOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}},{"in":"header","required":true,"name":"Accept-Language","schema":{"type":"string","default":"fr","enum":["fr","en","ar"]}}],"responses":{"200":{"description":""}},"tags":["Operator"]}},"/api/transaction/order":{"post":{"operationId":"TransactionController_order","parameters":[{"in":"header","required":true,"name":"Accept-Language","schema":{"type":"string","default":"fr","enum":["fr","en","ar"]}}],"requestBody":{"required":true,"description":"Json structure for Transaction object","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateTransactionDto"}}}},"responses":{"201":{"description":""}},"tags":["Transaction"]}},"/api/transaction":{"get":{"operationId":"TransactionController_getAll","parameters":[{"name":"lang","required":false,"in":"query","description":"The language for the request","schema":{"default":"en","example":"fr","enum":["fr","en","ar"],"type":"string"}},{"name":"ref","required":false,"in":"query","schema":{"example":"","type":"string"}},{"name":"customerAccountingCode","required":false,"in":"query","schema":{"example":"","type":"string"}},{"name":"productId","required":false,"in":"query","schema":{"example":"","type":"uuid"}},{"name":"fullName","required":false,"in":"query","schema":{"example":"","type":"string"}},{"name":"source","required":false,"in":"query","schema":{"example":"","type":"string"}},{"name":"page","required":false,"in":"query","schema":{"example":1,"type":"number"}},{"name":"pageSize","required":false,"in":"query","schema":{"example":10,"type":"number"}},{"name":"orderBy","required":false,"in":"query","schema":{"example":"createdAt","type":"string"}},{"name":"sortOrder","required":false,"in":"query","schema":{"example":"DESC","type":"string"}},{"in":"header","required":true,"name":"Accept-Language","schema":{"type":"string","default":"fr","enum":["fr","en","ar"]}}],"responses":{"200":{"description":""}},"tags":["Transaction"]}}},"info":{"title":"VOUCHER","description":"The vouchers API description","version":"1.0","contact":{}},"tags":[],"servers":[],"components":{"schemas":{"FiltersLogsDto":{"type":"object","properties":{"page":{"type":"number","example":1},"pageSize":{"type":"number","example":10},"orderBy":{"type":"string","example":"createdAt"},"sortOrder":{"type":"string","example":"DESC|ASC"},"lang":{"type":"string","description":"The language for the request","example":"fr","default":"en","enum":["fr","en","ar"]}}},"FiltersVoucherDto":{"type":"object","properties":{"lang":{"type":"string","description":"The language for the request","example":"fr","default":"en","enum":["fr","en","ar"]},"transactionId":{"type":"uuid","example":""},"documentId":{"type":"uuid","example":""},"productId":{"type":"uuid","example":""},"status":{"type":"number","example":1},"page":{"type":"number","example":1},"pageSize":{"type":"number","example":10},"orderBy":{"type":"string","example":"createdAt"},"sortOrder":{"type":"string","example":"DESC"}}},"FiltersProductDto":{"type":"object","properties":{"lang":{"type":"string","description":"The language for the request","example":"fr","default":"en","enum":["fr","en","ar"]},"productAccountingCode":{"type":"string","example":"1"},"facialId":{"type":"uuid","example":"...."},"operatorId":{"type":"uuid","example":"......"},"enabled":{"type":"boolean","example":true},"page":{"type":"number","example":1},"pageSize":{"type":"number","example":10},"orderBy":{"type":"string","example":"createdAt"},"sortOrder":{"type":"string","example":"DESC|ASC"}}},"PayloadById":{"type":"object","properties":{"lang":{"type":"string","description":"The language for the request","example":"fr","default":"en","enum":["fr","en","ar"]},"id":{"type":"uuid","example":true}},"required":["id"]},"CreateProductDto":{"type":"object","properties":{"lang":{"type":"string","description":"The language for the request","example":"fr","default":"en","enum":["fr","en","ar"]},"label":{"type":"string","example":"OOREDOO 5TND"},"productAccountingCode":{"type":"string","example":"1"},"facialId":{"type":"uuid","example":"1"},"operatorId":{"type":"uuid","example":"1"},"serviceId":{"type":"uuid","example":"..........."},"image":{"type":"string","example":"data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAA3"}},"required":["label","productAccountingCode","facialId","operatorId","serviceId"]},"UpdateProductDto":{"type":"object","properties":{"label":{"type":"string","example":"OOREDOO 5TND"},"productAccountingCode":{"type":"string","example":"1"},"facialId":{"type":"uuid","example":"1"},"operatorId":{"type":"uuid","example":"1"},"serviceId":{"type":"uuid","example":"..........."},"enabled":{"type":"boolean","example":true},"realQuantity":{"type":"number","example":true},"virtualQuantity":{"type":"number","example":true},"image":{"type":"string","example":"data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAA3"}},"required":["label","productAccountingCode","facialId","operatorId","serviceId"]},"UpdateProductDtoTcp":{"type":"object","properties":{"lang":{"type":"string","description":"The language for the request","example":"fr","default":"en","enum":["fr","en","ar"]},"id":{"type":"uuid","example":true},"updateProductDto":{"$ref":"#/components/schemas/UpdateProductDto"}},"required":["id","updateProductDto"]},"FiltersFacialDto":{"type":"object","properties":{"page":{"type":"number","example":1},"pageSize":{"type":"number","example":10},"orderBy":{"type":"string","example":"createdAt"},"sortOrder":{"type":"string","example":"DESC|ASC"},"lang":{"type":"string","description":"The language for the request","example":"fr","default":"en","enum":["fr","en","ar"]},"label":{"type":"string","example":"5 TND"},"value":{"type":"number","example":"1.7"},"enabled":{"type":"boolean","example":true}}},"CreateFacialDto":{"type":"object","properties":{"lang":{"type":"string","description":"The language for the request","example":"fr","default":"en","enum":["fr","en","ar"]},"label":{"type":"string","example":"5 TND"},"value":{"type":"number","example":"5.7"}},"required":["label","value"]},"UpdateFacialDtoTcp":{"type":"object","properties":{"lang":{"type":"string","description":"The language for the request","example":"fr","default":"en","enum":["fr","en","ar"]},"id":{"type":"uuid","example":true}},"required":["id"]},"UpdateFacialDto":{"type":"object","properties":{"label":{"type":"string","example":"5 TND"},"value":{"type":"number","example":"5.7"},"enabled":{"type":"boolean","example":true}},"required":["label","value"]},"FiltersOperatorDto":{"type":"object","properties":{"page":{"type":"number","example":1},"pageSize":{"type":"number","example":10},"orderBy":{"type":"string","example":"createdAt"},"sortOrder":{"type":"string","example":"DESC|ASC"},"lang":{"type":"string","description":"The language for the request","example":"fr","default":"en","enum":["fr","en","ar"]},"label":{"type":"string","example":"OOREDOO"},"enabled":{"type":"boolean","example":true}}},"CreateOperatorDto":{"type":"object","properties":{"lang":{"type":"string","description":"The language for the request","example":"fr","default":"en","enum":["fr","en","ar"]},"label":{"type":"string","example":"OOREDOO"},"description":{"type":"string","example":"*********"},"image":{"type":"string","example":"data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAA3"}},"required":["label","description","image"]},"UpdateOperatorDtoTcp":{"type":"object","properties":{"lang":{"type":"string","description":"The language for the request","example":"fr","default":"en","enum":["fr","en","ar"]},"id":{"type":"uuid","example":true}},"required":["id"]},"UpdateOperatorDto":{"type":"object","properties":{"label":{"type":"string","example":"OOREDOO"},"description":{"type":"string","example":"*********"},"enabled":{"type":"boolean","example":true},"image":{"type":"string","example":"data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAA3"}},"required":["label","description","image"]},"FiltersTransactionDto":{"type":"object","properties":{"lang":{"type":"string","description":"The language for the request","example":"fr","default":"en","enum":["fr","en","ar"]},"ref":{"type":"string","example":""},"customerAccountingCode":{"type":"string","example":""},"productId":{"type":"uuid","example":""},"fullName":{"type":"string","example":""},"source":{"type":"string","example":""},"page":{"type":"number","example":1},"pageSize":{"type":"number","example":10},"orderBy":{"type":"string","example":"createdAt"},"sortOrder":{"type":"string","example":"DESC"}}},"CreateTransactionDto":{"type":"object","properties":{"lang":{"type":"string","description":"The language for the request","example":"fr","default":"en","enum":["fr","en","ar"]},"customerAccountingCode":{"type":"string","example":""},"productId":{"type":"uuid","example":""},"fullName":{"type":"string","example":""},"customerId":{"type":"uuid","example":""},"quantity":{"type":"number"},"source":{"type":"string","example":""},"statementId":{"type":"uuid","example":""}}}}}}